feat: implement audio player, subtitle parsing, and intensive listening features
- Update article_segments migration and model to use decimal/float for start_time and end_time
- Add SRT and VTT parsing logic to ArticleTextProcessor
- Implement frontend audio player with sentence highlighting, click-to-seek, and loop listening in show.blade.php
- Add @yield('scripts') to layout for frontend scripts

Made-with: Cursor

+++ database/migrations/2026_03_22_110100_create_article_segments_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->unsignedInteger('sentence_index');
             $table->text('content_en');
             $table->text('content_cn')->nullable();
-            $table->integer('start_time')->nullable();
-            $table->integer('end_time')->nullable();
+            $table->decimal('start_time', 8, 3)->nullable();
+            $table->decimal('end_time', 8, 3)->nullable();
 
             $table->index(['article_id', 'paragraph_index'], 'idx_article_segment');
             $table->foreign('article_id')->references('article_id')->on('articles')->cascadeOnDelete();
